backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
requests==2.31.0
python-dotenv==1.0.0
pymysql==1.1.0
